feat: only install requirements if extension exists in manifest

+++ agents/scripts/install_deps_and_build.sh
@@ -60,7 +60,11 @@ install_python_requirements() {
   if [[ -d "ten_packages/extension" ]]; then
     for extension in ten_packages/extension/*; do
       if [[ -f "$extension/requirements.txt" ]]; then
-        pip install -r $extension/requirements.txt
+        local extension_name=$(basename $extension)
+        local exists_in_manifest=$(cat manifest.json | jq --arg ext_name "${extension_name}" 'any(.dependencies[].name == $ext_name ; .)' )
+        if [[ "$exists_in_manifest" == "true" ]]; then
+          pip install -r $extension/requirements.txt
+        fi
       fi
     done
   fi
